requirements.txt
chromadb
nltk
torch
PyYAML
ollama
PyMuPDF
sentence-transformers
tqdm
